+++ nautilus_trader/data/aggregation.pyx
@@ -124,7 +124,7 @@ cdef class BarBuilder:
         self._partial_set = True
         self.initialized = True
 
-    cpdef void update(self, double price, int64_t size, int64_t ts_event) except *:
+    cpdef void update(self, double price, double size, int64_t ts_event) except *:
         """
         Update the bar builder.
 
@@ -327,7 +327,7 @@ cdef class BarAggregator:
             ts_event=tick.ts_event,
         )
 
-    cdef void _apply_update(self, double price, int64_t size, int64_t ts_event) except *:
+    cdef void _apply_update(self, double price, double size, int64_t ts_event) except *:
         raise NotImplementedError("method must be implemented in the subclass")  # pragma: no cover
 
     cdef void _build_now_and_send(self) except *:
@@ -377,7 +377,7 @@ cdef class TickBarAggregator(BarAggregator):
             logger=logger,
         )
 
-    cdef void _apply_update(self, double price, int64_t size, int64_t ts_event) except *:
+    cdef void _apply_update(self, double price, double size, int64_t ts_event) except *:
         self._builder.update(price, size, ts_event)
 
         if self._builder.count == self.bar_type.spec.step:
@@ -422,7 +422,7 @@ cdef class VolumeBarAggregator(BarAggregator):
             logger=logger,
         )
 
-    cdef void _apply_update(self, double price, int64_t size, int64_t ts_event) except *:
+    cdef void _apply_update(self, double price, double size, int64_t ts_event) except *:
         size_update = size
 
         while size_update > 0:  # While there is size to apply
@@ -502,7 +502,7 @@ cdef class ValueBarAggregator(BarAggregator):
         """
         return self._cum_value
 
-    cdef void _apply_update(self, double price, int64_t size, int64_t ts_event) except *:
+    cdef void _apply_update(self, double price, double size, int64_t ts_event) except *:
         size_update = size
 
         while size_update > 0:  # While there is value to apply
@@ -698,7 +698,7 @@ cdef class TimeBarAggregator(BarAggregator):
 
         self._log.debug(f"Started timer {timer_name}.")
 
-    cdef void _apply_update(self, double price, int64_t size, int64_t ts_event) except *:
+    cdef void _apply_update(self, double price, double size, int64_t ts_event) except *:
         if self._clock.is_test_clock:
             if self.next_close_ns < ts_event:
                 # Build bar first, then update
